+++ src/Helpers/Importers/Product/Importer.php
@@ -3,10 +3,45 @@
 namespace Webkul\DAM\Helpers\Importers\Product;
 
 use Webkul\DAM\Models\Asset;
+use Webkul\DAM\Repositories\AssetRepository;
 use Webkul\DataTransfer\Helpers\Importers\Product\Importer as ProductImporter;
+use Webkul\DataTransfer\Helpers\Importers\Product\SKUStorage;
+use Webkul\Attribute\Repositories\AttributeFamilyRepository;
+use Webkul\Attribute\Repositories\AttributeOptionRepository;
+use Webkul\Attribute\Repositories\AttributeRepository;
+use Webkul\Category\Repositories\CategoryRepository;
+use Webkul\Core\Repositories\ChannelRepository;
+use Webkul\DataTransfer\Helpers\Importers\FieldProcessor;
+use Webkul\DataTransfer\Repositories\JobTrackBatchRepository;
+use Webkul\Product\Repositories\ProductRepository;
 
 class Importer extends ProductImporter
 {
+    public function __construct(
+        protected JobTrackBatchRepository $importBatchRepository,
+        protected AttributeFamilyRepository $attributeFamilyRepository,
+        protected AttributeRepository $attributeRepository,
+        protected AttributeOptionRepository $attributeOptionRepository,
+        protected CategoryRepository $categoryRepository,
+        protected ProductRepository $productRepository,
+        protected SKUStorage $skuStorage,
+        protected ChannelRepository $channelRepository,
+        protected FieldProcessor $fieldProcessor,
+        protected AssetRepository $assetRepository
+    ) {
+        parent::__construct(
+            $importBatchRepository,
+            $attributeFamilyRepository,
+            $attributeRepository,
+            $attributeOptionRepository,
+            $categoryRepository,
+            $productRepository,
+            $skuStorage,
+            $channelRepository,
+            $fieldProcessor
+        );
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -36,6 +71,24 @@ class Importer extends ProductImporter
             }
 
             if ($attribute->type === Asset::ASSET_ATTRIBUTE_TYPE) {
+                if (! empty($value)) {
+                    $values = explode(',', $value);
+
+                    $assets = [];
+                    foreach ($values as $value) {
+                        $asset = $this->assetRepository->findWhereIn('path', [trim($value)])->first();
+                        
+                        if ($asset) {
+                            $assets[] = $asset->id;
+                        }
+                    }
+
+                    if ($assets) {
+                        $value = implode(',', $assets);
+
+                        $attribute->setProductValue($value, $attributeValues, $rowData['channel'] ?? null, $rowData['locale'] ?? null);
+                    }
+                }
                 continue;
             }