+++ src/Http/Controllers/API/Asset/PropertyController.php
@@ -80,12 +80,13 @@ class PropertyController extends Controller
             ], 400);
         }
 
-        $property = $this->assetPropertyRepository->create(array_merge(request()->only([
-            'name',
-            'type',
-            'language',
-            'value',
-        ]), ['dam_asset_id' => $id]));
+        $property = $this->assetPropertyRepository->create([
+            'name'         => request('name'),
+            'type'         => request('type'),
+            'language'     => $locale->id,
+            'value'        => request('value'),
+            'dam_asset_id' => $id,
+        ]);
 
         return response()->json([
             'success'  => true,
@@ -99,10 +100,6 @@ class PropertyController extends Controller
      */
     public function update(Request $request, int $id): JsonResponse
     {
-        $this->validate($request, [
-            'value' => 'required',
-        ]);
-
         $property = $this->assetPropertyRepository->find($id);
 
         if (! $property) {
@@ -112,8 +109,13 @@ class PropertyController extends Controller
             ], 404);
         }
 
+        $this->validate($request, [
+            'name'  => 'required|min:3|max:100|unique:dam_asset_properties,name,'.$id.',id,dam_asset_id,'.$property->dam_asset_id,
+            'value' => 'required',
+        ]);
+
         try {
-            $updatedProperty = $this->assetPropertyRepository->update($request->only(['value']), $id);
+            $updatedProperty = $this->assetPropertyRepository->update($request->only(['name', 'value']), $id);
 
             return response()->json([
                 'success'  => true,